+++ CHANGELOG.md
@@ -17,7 +17,7 @@ History:
 ### v1.1.5 -> v1.x.x (2010/MM/DD)
 
  - [+] 单元测试框架改用 Jasmine 20% [tests](http://kissyteam.github.com/kissy/tests/)
- - [+] imagezoom 增加 Zazzle 内部放大效果 [demo](todo)
+ - [+] imagezoom 增加 Zazzle 内部放大效果 [demo](http://kissyui.com/blog/2010/10/imagezoom-inner/)
  - [*] anim 增加原生 CSS3 Transitions 支持 [原理](http://lifesinger.org/blog/2010/09/anim-using-css3-transitions/)
  - [+] 组件开发规范整理 20%
  - [*] 组件开发流程整理 80%