+++ CHANGELOG.md
@@ -16,6 +16,7 @@ History:
 
 ### v1.1.5 -> v1.x.x (2010/MM/DD)
 
+ - [+] 增加了 append/appendTo/prepend/prependTo [尚缺文档]
  - [+] 采用 Jasmine 构建所有正式组件的测试用例 [Understanding JavaScript Testing](http://kissyui.com/blog/2010/10/understanding-javascript-testing/)
  - [+] gallery 增加 Sliding Label demo
  - [+] imagezoom 增加 Zazzle 内部放大效果 [demo](http://kissyteam.github.com/kissy/src/imagezoom/test-inner.html)