+++ CHANGELOG.md
@@ -16,11 +16,11 @@ History:
 
 ### v1.1.5 -> v1.x.x (2010/MM/DD)
 
- - [*] 升级 sizzle 到 v1.4.4
+ - [+] 2010/11/30 发布内部版本 v1.1.6
  - [+] 增加 ajax 模块 [文档](http://kissyteam.github.com/kissy/docs/ajax/index.html)
  - [!] 将 S.one/all(selector).on(type, function(ev) { ev.target/currentTarget 由原来的裸节点变成对应 Node })
  - [!] 将 S.all(selector).on(type, function() { this 由 NodeList 变成对应 Node })
- - [+] 增加了 append/appendTo/prepend/prependTo [尚缺文档]
+ - [+] dom 增加 append/appendTo/prepend/prependTo [尚缺文档]
  - [+] 引入 Jasmine [Understanding JavaScript Testing](http://kissyui.com/blog/2010/10/understanding-javascript-testing/)
  - [+] Gallery 增加 Sliding Label [demo](http://kissyteam.github.com/kissy-gallery/slidinglabels/demo.html)
  - [+] imagezoom 增加 Zazzle 内部放大效果 [demo](http://kissyteam.github.com/kissy/src/imagezoom/test-inner.html)